requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.2.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
pyodbc==5.1.0
python-dotenv==1.0.1
gunicorn
psycopg2-binary
